performance: card performance optimization & SEO setting
