+++ README.md
@@ -40,7 +40,7 @@
 
 - 🍰 内置 `800+` 优质网站。
 - 🍰 支持 [码云 Gitee](https://gitee.com/xiejiahe/nav)
-- 🍰 支持 SEO。
+- 🍰 支持 SEO 搜索引擎。
 - 🍰 完全纯静态, 提供自动化部署功能。
 - 🍰 三叉树分类、结构清晰、分类清晰。
 - 🍰 支持一个网站关联多个网址
@@ -107,7 +107,7 @@
 
 ## 书签导入
 
-支持从 Chrome 书签导入（WebKit 内核应该都是支持的~），会自动检测满足三级分类的导航，其他一律设为未分类：
+自动检测满足三级分类导航，其他一律设为未分类：
 
 ![](https://raw.githubusercontent.com/xjh22222228/public/gh-pages/nav/import.png)
 
@@ -125,7 +125,7 @@ git merge upstream/main --allow-unrelated-histories
 git push
 
 # 或者执行
-npm run fetch
+npm run update
 ```
 
 ## 更新日志
