+++ package.json
@@ -18,7 +18,7 @@
     "build-gh-pages": "npm run setup && ng build --base-href ./ --index src/index.html",
     "build": "npm run setup && ng build --index src/index.html",
     "watch": "ng build --watch --configuration development",
-    "fetch": "git pull && (git remote add upstream https://github.com/xjh22222228/nav.git || true) && git fetch upstream main && git merge upstream/main --allow-unrelated-histories && git push"
+    "update": "git pull && (git remote add upstream https://github.com/xjh22222228/nav.git || true) && git fetch upstream main && git merge upstream/main --allow-unrelated-histories && git push"
   },
   "private": true,
   "dependencies": {
