+++ src/app/app-routing.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core'
 import { RouterModule, Routes } from '@angular/router'
 import config from '../../nav.config'
 import { settings } from 'src/store'
-import LightComponent from '../view/index/light/index.component'
-import SuperComponent from '../view/index/super/index.component'
-import SimComponent from '../view/index/sim/index.component'
+import LightComponent from '../view/light/index.component'
+import SuperComponent from '../view/super/index.component'
+import SimComponent from '../view/sim/index.component'
 import SystemComponent from '../view/system/index.component'
 import SystemInfoComponent from '../view/system/info/index.component'
 import SystemBookmarkComponent from '../view/system/bookmark/index.component'
@@ -14,8 +14,8 @@ import SystemSearchComponent from '../view/system/search/index.component'
 import SystemSettingComponent from '../view/system/setting/index.component'
 import SystemWebComponent from '../view/system/web/index.component'
 import SystemAngleMarkComponent from '../view/system/angle-mark/index.component'
-import SideComponent from '../view/index/side/index.component'
-import ShortcutComponent from '../view/index/shortcut/index.component'
+import SideComponent from '../view/side/index.component'
+import ShortcutComponent from '../view/shortcut/index.component'
 import WebpComponent from '../view/app/default/app.component'
 
 const routes: Routes = [
