+++ src/app/app.module.ts
@@ -54,9 +54,9 @@ import { CardComponent } from '../components/card/index.component'
 import { MoveWebComponent } from '../components/move-web/index.component'
 import { WebMoreMenuComponent } from '../components/web-more-menu/index.component'
 
-import LightComponent from '../view/index/light/index.component'
-import SuperComponent from '../view/index/super/index.component'
-import SimComponent from '../view/index/sim/index.component'
+import LightComponent from '../view/light/index.component'
+import SuperComponent from '../view/super/index.component'
+import SimComponent from '../view/sim/index.component'
 import SystemComponent from '../view/system/index.component'
 import SystemInfoComponent from '../view/system/info/index.component'
 import SystemBookmarkComponent from '../view/system/bookmark/index.component'
@@ -66,8 +66,8 @@ import SystemSearchComponent from '../view/system/search/index.component'
 import SystemSettingComponent from '../view/system/setting/index.component'
 import SystemWebComponent from '../view/system/web/index.component'
 import SystemAngleMarkComponent from '../view/system/angle-mark/index.component'
-import SideComponent from '../view/index/side/index.component'
-import ShortcutComponent from '../view/index/shortcut/index.component'
+import SideComponent from '../view/side/index.component'
+import ShortcutComponent from '../view/shortcut/index.component'
 import WebpComponent from '../view/app/default/app.component'
 
 import {
