+++ src/components/card/index.component.scss
@@ -1,135 +1,185 @@
 .container {
   position: relative;
   height: 100%;
-
-  .rate {
-    margin-bottom: 10px;
-    margin-top: -10px;
-  }
-
   .desc {
     color: rgba(0, 0, 0, 0.45);
     word-break: break-all;
     white-space: pre-wrap;
   }
-
   .title {
     white-space: normal;
     word-break: break-all;
   }
+}
 
-  &.example2 {
-    background-color: #fff;
-    padding: 8px 8px;
-    border-radius: 8px;
-    cursor: pointer;
-    box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1),
-      0 1px 2px -1px rgba(0, 0, 0, 0.1);
-    border: 1px solid transparent;
-    transition: all 0.15s linear;
+.standard {
+  box-shadow: 0 1px 5px rgba(0, 0, 0, 0.1);
+  $padding: 8px;
+  border-radius: 4px;
+  border: 1px solid #eee;
+  display: flex;
+  flex-direction: column;
+  .rate {
+    margin-bottom: 5px;
+    padding: 0 $padding;
+  }
+  .middle {
+    flex: 1;
+  }
+  .top {
     display: flex;
-    color: #000;
-    &:hover {
-      border: 1px solid #78b7fa;
-      .title {
-        color: #007bff;
-      }
-    }
-    .right-box {
-      flex: 1;
-      width: 0;
-      margin-left: 12px;
-    }
+    align-items: center;
+    padding: 0 $padding;
     .title {
-      display: flex;
-      align-items: center;
-    }
-    .desc {
-      margin: 0;
-      font-size: 12px;
+      margin-left: $padding;
+      font-weight: 500;
+      color: #000;
     }
   }
-
-  &.column2 {
-    background-color: #fff;
-    padding: 8px 12px;
-    border-bottom: 1px solid #eee;
-    transition: all 0.15s linear;
-    color: #000;
-    display: block;
+  .desc {
+    padding: $padding $padding 0;
+    margin-bottom: 5px;
+  }
+  .tagbox {
+    display: flex;
+    padding: 0 $padding;
+    flex-wrap: wrap;
+  }
+  .tag-item {
+    margin: 6px 6px 0 0;
+    padding: 0 6px;
+    border-radius: 2px;
+    font-size: 12px;
+    display: flex;
+    align-items: center;
+    transition: all 0.1s linear;
+    color: #fff;
+    a {
+      color: currentColor;
+    }
     &:hover {
-      background-color: #f2f3f5;
-      .action-btn {
-        display: block;
-      }
+      opacity: 0.8;
     }
-    .title-bar {
-      display: flex;
-      align-items: center;
-    }
-    .right-box {
-      margin-top: 6px;
-    }
-    .title {
-      font-weight: bold;
-      margin-left: 10px;
-      font-size: 16px;
-    }
-    .desc {
-      margin: 0;
-      font-size: 12px;
-    }
-    .bottom-bar {
-      display: flex;
-      margin-top: 4px;
-    }
-    .tagbox {
-      display: flex;
+  }
+  .actionbar {
+    display: flex;
+    border-top: 1px solid #eee;
+    padding: 8px 0;
+    margin-top: 10px;
+    .action-item {
       flex: 1;
-      justify-content: flex-start;
-      align-items: center;
-      a {
-        color: #fff;
-      }
-    }
-    .tag-item {
-      margin-right: 6px;
-      padding: 0 6px;
-      border-radius: 2px;
-      font-size: 12px;
-      display: flex;
-      align-items: center;
-      transition: all 0.1s linear;
-      &:hover {
-        opacity: 0.8;
-      }
-    }
-    .rate2 {
-      color: #8a919f;
-      margin-right: 10px;
-    }
-    .action-btn {
-      margin-right: 4px;
-      color: #8a919f;
-      font-size: 12px;
-      display: none;
+      text-align: center;
+      font-size: 16px;
+      color: #999;
       cursor: pointer;
       &:hover {
-        color: #666;
+        color: #1890ff;
+      }
+      &:not(:nth-last-child(1)) {
+        border-right: 1px solid #eee;
       }
     }
   }
 }
 
-.add-wrapper {
+.example2 {
+  background-color: #fff;
+  padding: 8px 8px;
+  border-radius: 8px;
   cursor: pointer;
+  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px -1px rgba(0, 0, 0, 0.1);
+  border: 1px solid transparent;
+  transition: all 0.15s linear;
   display: flex;
-  align-items: center;
-  justify-content: center;
+  color: #000;
+  &:hover {
+    border: 1px solid #78b7fa;
+    .title {
+      color: #007bff;
+    }
+  }
+  .right-box {
+    flex: 1;
+    width: 0;
+    margin-left: 12px;
+  }
+  .title {
+    display: flex;
+    align-items: center;
+  }
+  .desc {
+    margin: 0;
+    font-size: 12px;
+  }
+}
 
-  .add-icon {
-    font-size: 50px;
-    color: #ccc;
+.column2 {
+  background-color: #fff;
+  padding: 8px 12px;
+  border-bottom: 1px solid #eee;
+  transition: all 0.15s linear;
+  color: #000;
+  display: block;
+  &:hover {
+    background-color: #f2f3f5;
+    .action-btn {
+      display: block;
+    }
+  }
+  .title-bar {
+    display: flex;
+    align-items: center;
+  }
+  .right-box {
+    margin-top: 6px;
+  }
+  .title {
+    font-weight: bold;
+    margin-left: 10px;
+    font-size: 16px;
+  }
+  .desc {
+    margin: 0;
+    font-size: 12px;
+  }
+  .bottom-bar {
+    display: flex;
+    margin-top: 4px;
+  }
+  .tagbox {
+    display: flex;
+    flex: 1;
+    justify-content: flex-start;
+    align-items: center;
+    a {
+      color: #fff;
+    }
+  }
+  .tag-item {
+    margin-right: 6px;
+    padding: 0 6px;
+    border-radius: 2px;
+    font-size: 12px;
+    display: flex;
+    align-items: center;
+    transition: all 0.1s linear;
+    &:hover {
+      opacity: 0.8;
+    }
+  }
+  .rate2 {
+    color: #8a919f;
+    margin-right: 10px;
+  }
+  .action-btn {
+    margin-right: 4px;
+    color: #8a919f;
+    font-size: 12px;
+    display: none;
+    cursor: pointer;
+    &:hover {
+      color: #666;
+    }
   }
 }
