+++ src/view/light/index.component.scss
