+++ src/view/shortcut/index.component.scss
