+++ src/view/shortcut/index.component.ts
@@ -5,18 +5,15 @@ import { isDark as isDarkFn } from 'src/utils'
 import { settings } from 'src/store'
 import event from 'src/utils/mitt'
 
-const { title, shortcutThemeShowWeather, shortcutThemeImages } = settings
-
 @Component({
   selector: 'app-shortcut',
   templateUrl: './index.component.html',
   styleUrls: ['./index.component.scss'],
 })
 export default class ShortcutComponent {
-  title = title
+  settings = settings
   isDark: boolean = isDarkFn()
-  shortcutThemeShowWeather = shortcutThemeShowWeather
-  shortcutThemeImage = shortcutThemeImages[0]['src']
+  shortcutThemeImage = settings.shortcutThemeImages?.[0]?.['src']
 
   constructor() {
     event.on('EVENT_DARK', (isDark: any) => {
