+++ src/view/side/index.component.scss
