+++ src/view/sim/index.component.ts
@@ -1,5 +1,4 @@
 // Copyright @ 2018-present xiejiahe. All rights reserved. MIT license.
-// See https://github.com/xjh22222228/nav
 
 import { Component } from '@angular/core'
 import { Router, ActivatedRoute } from '@angular/router'
@@ -10,27 +9,30 @@ import {
   setWebsiteList,
   toggleCollapseAll,
   matchCurrentList,
+  getOverIndex,
 } from 'src/utils'
 import { isLogin } from 'src/utils/user'
 import { websiteList } from 'src/store'
-import { settings, searchEngineList } from 'src/store'
+import { settings, internal } from 'src/store'
 
 @Component({
-  selector: 'app-side',
+  selector: 'app-sim',
   templateUrl: './index.component.html',
   styleUrls: ['./index.component.scss'],
 })
-export default class SideComponent {
-  LOGO_CDN = settings.favicon
+export default class SimComponent {
   websiteList: INavProps[] = websiteList
   currentList: INavThreeProp[] = []
   id: number = 0
   page: number = 0
-  title: string = settings.title.trim().split(/\s/)[0]
-  searchEngineList = searchEngineList
-  isLogin = isLogin
   settings = settings
+  description: string = settings.simThemeDesc.replace(
+    '${total}',
+    String(isLogin ? internal.loginViewCount : internal.userViewCount)
+  )
+  isLogin = isLogin
   sliceMax = 1
+  overIndex = Number.MAX_SAFE_INTEGER
 
   constructor(private router: Router, private activatedRoute: ActivatedRoute) {}
 
@@ -48,44 +50,45 @@ export default class SideComponent {
       }
       setTimeout(() => {
         this.sliceMax = Number.MAX_SAFE_INTEGER
-      }, 25)
+      }, 100)
     })
   }
 
-  handleSidebarNav(page: any, id: any) {
-    this.websiteList[page].id = id
+  ngOnDestroy() {}
+
+  ngAfterViewInit() {
+    if (this.settings.simOverType === 'ellipsis') {
+      queueMicrotask(() => {
+        const overIndex = getOverIndex('.top-nav .over-item')
+        if (this.overIndex === overIndex) {
+          return
+        }
+        this.overIndex = overIndex
+      })
+    }
+  }
+
+  handleSidebarNav(index: number) {
+    const { page } = queryString()
+    this.websiteList[page].id = index
     this.router.navigate([this.router.url.split('?')[0]], {
       queryParams: {
         page,
-        id,
+        id: index,
+        _: Date.now(),
       },
     })
-    this.handlePositionTop()
   }
 
-  handlePositionTop() {
-    setTimeout(() => {
-      const el = document.querySelector('.search-header') as HTMLDivElement
-      if (el) {
-        const h = el.offsetHeight
-        window.scroll({
-          top: h,
-          left: 0,
-          behavior: 'smooth',
-        })
-      }
-    }, 30)
-  }
-
-  openMenu(item: any, index: number) {
-    this.websiteList.forEach((data, idx) => {
-      if (idx === index) {
-        data.collapsed = !data.collapsed
-      } else {
-        data.collapsed = false
-      }
+  handleCilckTopNav(idx: number) {
+    const id = this.websiteList[idx].id || 0
+    this.router.navigate([this.router.url.split('?')[0]], {
+      queryParams: {
+        page: idx,
+        id,
+        _: Date.now(),
+      },
     })
-    setWebsiteList(this.websiteList)
   }
 
   onCollapse = (item: any, index: number) => {
@@ -94,10 +97,8 @@ export default class SideComponent {
     setWebsiteList(this.websiteList)
   }
 
-  onCollapseAll = (e: Event) => {
-    e?.stopPropagation()
+  onCollapseAll = () => {
     toggleCollapseAll(this.websiteList)
-    this.handlePositionTop()
   }
 
   collapsed() {
@@ -106,5 +107,13 @@ export default class SideComponent {
     } catch (error) {
       return false
     }
+  }
+
+  trackByItem(a: any, item: any) {
+    return item.title
+  }
+
+  trackByItemWeb(a: any, item: any) {
+    return item.id
   }
 }
