+++ src/view/super/index.component.html
@@ -8,7 +8,11 @@
 
       <div class="topnav dark-scrollbar">
         <div
-          *ngFor="let item of websiteList.slice(0, overIndex); index as i"
+          *ngFor="
+            let item of websiteList.slice(0, overIndex);
+            index as i;
+            trackBy: trackByItem
+          "
           (click)="handleCilckTopMenu(i)"
           [class.active]="page === i"
           [class.dark-text-active]="page === i"
@@ -39,7 +43,11 @@
         [class.active]="id === i"
         [class.dark-item-active]="id === i"
         (click)="handleSidebarNav(i)"
-        *ngFor="let item of websiteList[page].nav; index as i"
+        *ngFor="
+          let item of websiteList[page].nav;
+          index as i;
+          trackBy: trackByItem
+        "
       >
         <img *ngIf="item.icon" class="sideicon" [src]="item.icon" />
         <span class="ellipsis">{{ item.title }}</span>
@@ -55,7 +63,7 @@
         <div class="bar dark-scrollbar">
           <div
             class="bar-item"
-            *ngFor="let item of currentList; index as i"
+            *ngFor="let item of currentList; index as i; trackBy: trackByItem"
             [class.active]="selectedIndex === i"
             [class.none]="!item.title"
             (click)="handleCheckThree(i)"
@@ -67,7 +75,11 @@
 
       <div *ngIf="currentList[selectedIndex]?.nav; else noData">
         <app-card
-          *ngFor="let el of currentList[selectedIndex].nav; index as j"
+          *ngFor="
+            let el of currentList[selectedIndex].nav;
+            index as j;
+            trackBy: trackByItemWeb
+          "
           [indexs]="[page, id, selectedIndex, j]"
           [cardStyle]="settings.superCardStyle"
           [dataSource]="el"
