+++ src/view/system/index.component.html
@@ -1,6 +1,6 @@
 <nz-layout class="system-layout" *ngIf="isLogin">
   <nz-layout>
-    <nz-sider nzWidth="150px" nzTheme="light">
+    <nz-sider class="sidebar" nzWidth="150px" nzTheme="light">
       <ul nz-menu nzMode="inline" class="sider-menu">
         <li nz-submenu nzOpen nzTitle="超级系统">
           <ul>
